@@ -1,6 +1,6 @@
 {
 	"name": "brew-guide",
-	"version": "2.0.2",
+	"version": "2.0.3",
 	"private": true,
 	"scripts": {
 		"dev": "next dev",
@@ -9,24 +9,25 @@
 		"lint": "next lint"
 	},
 	"dependencies": {
+		"@tailwindcss/postcss": "^4.0.13",
 		"@vercel/analytics": "^1.5.0",
-		"framer-motion": "^11.18.2",
-		"geist": "^1.2.1",
-		"next": "^15.2.1",
+		"framer-motion": "^12.5.0",
+		"geist": "^1.3.1",
+		"next": "^15.2.2",
 		"next-pwa": "^5.6.0",
-		"next-themes": "^0.4.4",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0"
+		"next-themes": "^0.4.6",
+		"react": "^19.0.0",
+		"react-dom": "^19.0.0"
 	},
 	"devDependencies": {
-		"@types/node": "^20.11.16",
-		"@types/react": "^18.2.52",
-		"@types/react-dom": "^18.2.18",
-		"autoprefixer": "^10.4.17",
-		"eslint": "^8.56.0",
-		"eslint-config-next": "15.2.1",
-		"postcss": "^8.4.35",
-		"tailwindcss": "^3.4.1",
-		"typescript": "^5.3.3"
+		"@types/node": "^22.13.10",
+		"@types/react": "^19.0.10",
+		"@types/react-dom": "^19.0.4",
+		"autoprefixer": "^10.4.21",
+		"eslint": "^9.22.0",
+		"eslint-config-next": "15.2.2",
+		"postcss": "^8.5.3",
+		"tailwindcss": "^4.0.13",
+		"typescript": "^5.8.2"
 	}
 }
