@@ -1,5 +1,5 @@
 // App Version
-export const APP_VERSION = "2.0.2";
+export const APP_VERSION = "2.0.3";
 
 // Types
 export interface Stage {
